+++ docs/release_notes/v0.15.0.rst
@@ -11,7 +11,7 @@ c. Avoidance of unexpected edge cases when more than one model is instantiated i
 However, there are several breaking changes for model developers. The data handling tutorial goes over these
 changes in detail.
 
-This refactor is centered around the new :class:`~scvi.data.anndata.AnnDataManager` class which orchestrates any data processing necessary
+This refactor is centered around the new :class:`~scvi.data.AnnDataManager` class which orchestrates any data processing necessary
 for scvi-tools and stores necessary information, rather than adding additional fields to the AnnData input.
 
 .. figure:: figures/anndata_manager_schematic.svg
@@ -23,8 +23,8 @@ for scvi-tools and stores necessary information, rather than adding additional f
 
 Changes
 ~~~~~~~
-- Major refactor to data handling strategy with the introduction of :class:`~scvi.data.anndata.AnnDataManager` (`#1237`_).
-- Prevent clobbering between models using the same AnnData object with model instance specific :class:`~scvi.data.anndata.AnnDataManager` mappings (`#1342`_).
+- Major refactor to data handling strategy with the introduction of :class:`~scvi.data.AnnDataManager` (`#1237`_).
+- Prevent clobbering between models using the same AnnData object with model instance specific :class:`~scvi.data.AnnDataManager` mappings (`#1342`_).
 - Add ``size_factor_key`` to :class:`~scvi.model.SCVI`, :class:`~scvi.model.MULTIVI`, :class:`~scvi.model.SCANVI`, and :class:`~scvi.model.TOTALVI` (`#1334`_).
 - Add references to the scvi-tools journal publication to the README (`#1338`_, `#1339`_).
 
@@ -33,18 +33,18 @@ Breaking changes
 - The keyword argument ``run_setup_anndata`` has been removed from built-in datasets since there is no longer a model-agnostic ``setup_anndata`` method (`#1237`_).
 - The function ``scvi.model._metrics.clustering_scores`` has been removed due to incompatbility with new data handling (`#1237`_).
 - :class:`~scvi.model.SCANVI` now takes ``unlabeled_category`` as an argument to :meth:`~scvi.model.SCANVI.setup_anndata` rather than on initialization (`#1237`_).
-- ``setup_anndata`` is now a class method on model classes and requires specific function calls to ensure proper :class:`~scvi.data.anndata.AnnDataManager` setup and model save/load.
+- ``setup_anndata`` is now a class method on model classes and requires specific function calls to ensure proper :class:`~scvi.data.AnnDataManager` setup and model save/load.
   Any model inheriting from :class:`~scvi.model.base.BaseModelClass` will need to re-implement this method (`#1237`_).
     - To adapt existing custom models to v0.15.0, one can references the guidelines below. For some examples of how this was done for the existing models in the codebase, please reference the following PRs: (`#1301`_, `#1302`_).
         1. ``scvi._CONSTANTS`` has been changed to ``scvi.REGISTRY_KEYS``.
         2. ``setup_anndata()`` functions are now class functions and follow a specific structure. Please refer to :meth:`~scvi.model.SCVI.setup_anndata` for an example.
-        3. ``scvi.data.get_from_registry()`` has been removed. This method can be replaced by :meth:`scvi.data.anndata.AnnDataManager.get_from_registry`.
-        4. The setup dict stored directly on the AnnData object, ``adata["_scvi"]``, has been deprecated. Instead, this information now lives in :attr:`scvi.data.anndata.AnnDataManager.registry`.
-            - The data registry can be accessed at :attr:`scvi.data.anndata.AnnDataManager.data_registry`.
-            - Summary stats can be accessed at :attr:`scvi.data.anndata.AnnDataManager.summary_stats`.
-            - Any field-specific information (e.g. ``adata.obs["categorical_mappings"]``) now lives in field-specific state registries. These can be retrieved via the function :meth:`~scvi.data.anndata.AnnDataManager.get_state_registry`.
+        3. ``scvi.data.get_from_registry()`` has been removed. This method can be replaced by :meth:`scvi.data.AnnDataManager.get_from_registry`.
+        4. The setup dict stored directly on the AnnData object, ``adata["_scvi"]``, has been deprecated. Instead, this information now lives in :attr:`scvi.data.AnnDataManager.registry`.
+            - The data registry can be accessed at :attr:`scvi.data.AnnDataManager.data_registry`.
+            - Summary stats can be accessed at :attr:`scvi.data.AnnDataManager.summary_stats`.
+            - Any field-specific information (e.g. ``adata.obs["categorical_mappings"]``) now lives in field-specific state registries. These can be retrieved via the function :meth:`~scvi.data.AnnDataManager.get_state_registry`.
         5. ``register_tensor_from_anndata()`` has been removed. To register tensors with no relevant ``AnnDataField`` subclass, create a new
-           a new subclass of :class:`~scvi.data.anndata.fields.BaseAnnDataField` and add it to appropriate model's ``setup_anndata()`` function.
+           a new subclass of :class:`~scvi.data.fields.BaseAnnDataField` and add it to appropriate model's ``setup_anndata()`` function.
 
 Contributors
 ~~~~~~~~~~~~
