+++ scvi/data/_compat.py
@@ -109,11 +109,11 @@ def manager_from_setup_dict(
     cls, adata: AnnData, setup_dict: dict, **transfer_kwargs
 ) -> AnnDataManager:
     """
-    Creates an :class:`~scvi.data.anndata.AnnDataManager` given only a scvi-tools setup dictionary.
+    Creates an :class:`~scvi.data.AnnDataManager` given only a scvi-tools setup dictionary.
 
     Only to be used for backwards compatibility when loading setup dictionaries for models.
-    Infers the AnnDataField instances used to define the :class:`~scvi.data.anndata.AnnDataManager` instance,
-    then uses the :meth:`~scvi.data.anndata.AnnDataManager.transfer_setup` method to register the new AnnData object.
+    Infers the AnnDataField instances used to define the :class:`~scvi.data.AnnDataManager` instance,
+    then uses the :meth:`~scvi.data.AnnDataManager.transfer_setup` method to register the new AnnData object.
 
     Parameters
     ----------
