--- scvi/data/anndata/__init__.py
@@ -1,5 +1,0 @@
-from ._manager import AnnDataManager
-
-__all__ = [
-    "AnnDataManager",
-]
