+++ scvi/data/fields/_obsm_field.py
@@ -8,9 +8,11 @@ import rich
 from anndata import AnnData
 from pandas.api.types import CategoricalDtype
 
-from scvi.data._utils import _check_nonnegative_integers
-from scvi.data.anndata import _constants
-from scvi.data.anndata._utils import _verify_and_correct_data_format
+from scvi.data import _constants
+from scvi.data._utils import (
+    _check_nonnegative_integers,
+    _verify_and_correct_data_format,
+)
 
 from ._base_field import BaseAnnDataField
