+++ scvi/data/fields/_protein.py
