+++ scvi/dataloaders/_concat_dataloader.py
@@ -4,7 +4,7 @@ from typing import List, Optional, Union
 import numpy as np
 from torch.utils.data import DataLoader
 
-from scvi.data.anndata import AnnDataManager
+from scvi.data import AnnDataManager
 
 from ._ann_dataloader import AnnDataLoader
 
@@ -16,7 +16,7 @@ class ConcatDataLoader(DataLoader):
     Parameters
     ----------
     adata_manager
-        :class:`~scvi.data.anndata.AnnDataManager` object that has been created via ``setup_anndata``.
+        :class:`~scvi.data.AnnDataManager` object that has been created via ``setup_anndata``.
     indices_list
         List where each element is a list of indices in the adata to load
     shuffle
