+++ scvi/external/cellassign/_model.py
@@ -8,8 +8,8 @@ from anndata import AnnData
 from pytorch_lightning.callbacks import Callback
 
 from scvi import REGISTRY_KEYS
-from scvi.data.anndata import AnnDataManager
-from scvi.data.anndata.fields import (
+from scvi.data import AnnDataManager
+from scvi.data.fields import (
     CategoricalJointObsField,
     CategoricalObsField,
     LayerField,
