+++ scvi/model/_multivi.py
@@ -12,8 +12,8 @@ from torch.distributions import Normal
 from scvi import REGISTRY_KEYS
 from scvi._compat import Literal
 from scvi._utils import _doc_params
-from scvi.data.anndata import AnnDataManager
-from scvi.data.anndata.fields import (
+from scvi.data import AnnDataManager
+from scvi.data.fields import (
     CategoricalJointObsField,
     CategoricalObsField,
     LayerField,
