+++ scvi/model/_scanvi.py
@@ -9,9 +9,9 @@ from anndata import AnnData
 
 from scvi import REGISTRY_KEYS
 from scvi._compat import Literal
-from scvi.data.anndata import AnnDataManager
-from scvi.data.anndata._constants import _SETUP_ARGS_KEY
-from scvi.data.anndata.fields import (
+from scvi.data import AnnDataManager
+from scvi.data._constants import _SETUP_ARGS_KEY
+from scvi.data.fields import (
     CategoricalJointObsField,
     CategoricalObsField,
     LabelsWithUnlabeledObsField,
