+++ scvi/model/_utils.py
@@ -8,7 +8,7 @@ import scipy.sparse as sp_sparse
 import torch
 
 from scvi import REGISTRY_KEYS
-from scvi.data.anndata import AnnDataManager
+from scvi.data import AnnDataManager
 
 logger = logging.getLogger(__name__)
 
@@ -69,7 +69,7 @@ def scrna_raw_counts_properties(
     Parameters
     ----------
     adata_manager
-        :class:`~scvi.data.anndata.AnnDataManager` object setup with :class:`~scvi.model.SCVI`.
+        :class:`~scvi.data.AnnDataManager` object setup with :class:`~scvi.model.SCVI`.
     idx1
         subset of indices describing the first population.
     idx2
@@ -136,7 +136,7 @@ def cite_seq_raw_counts_properties(
     Parameters
     ----------
     adata_manager
-        :class:`~scvi.data.anndata.AnnDataManager` object setup with :class:`~scvi.model.TOTALVI`.
+        :class:`~scvi.data.AnnDataManager` object setup with :class:`~scvi.model.TOTALVI`.
     idx1
         subset of indices describing the first population.
     idx2
@@ -181,7 +181,7 @@ def scatac_raw_counts_properties(
     Parameters
     ----------
     adata_manager
-        :class:`~scvi.data.anndata.AnnDataManager` object setup with :class:`~scvi.model.SCVI`.
+        :class:`~scvi.data.AnnDataManager` object setup with :class:`~scvi.model.SCVI`.
     idx1
         subset of indices describing the first population.
     idx2
@@ -236,7 +236,7 @@ def _init_library_size(
     Parameters
     ----------
     adata_manager
-        :class:`~scvi.data.anndata.AnnDataManager` object setup with :class:`~scvi.model.SCVI`.
+        :class:`~scvi.data.AnnDataManager` object setup with :class:`~scvi.model.SCVI`.
     n_batch
         Number of batches.
